Update README.md
I was looking at this documentation while doing the change logs today and reviewed it.

Please double check line 153 "When the
COUNT-based window becomes full, the oldest (last) data sample is removed to
free up space for an incoming sample"

I removed the parenthetical addition "(last)" because it seems to me that it is contradictory. The oldest sample should be the FIRST sample, while the last sample should be the newest.